+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## v2.0.0-beta.17
+
+### Features
+
+Updates the styles of `Button` and `Link` components
+
+### Breaking changes
+
+- `Button` and `Link` now have 4 sizes defined by the `size` prop (the `small` prop is removed)
+- `Button` variant 'subtle' is removed and replaced by 'quiet' and 'ghost' variants
+
 ## v2.0.0-beta.16
 
 ### Features